+++ src/main.rs
@@ -1,3 +1,4 @@
+mod package;
 mod parse;
 mod stability;
