+++ src/parse.rs
@@ -1,31 +1,9 @@
+use crate::package::MavenPackage;
 use indicatif::{ProgressBar, ProgressStyle};
 use log::debug;
 use roxmltree::{Document, NodeType};
 use semver::Version;
-use std::{
-    collections::HashMap,
-    convert::TryInto,
-    fmt::{Debug, Display, Formatter, Result},
-};
-
-/// Struct that represents a Maven package
-#[derive(Debug)]
-pub(crate) struct MavenPackage {
-    pub(crate) group_id: String,
-    pub(crate) artifact_id: String,
-    pub(crate) latest_version: String,
-    pub(crate) all_versions: Vec<Version>,
-}
-
-impl Display for MavenPackage {
-    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
-        write!(
-            f,
-            "{}:{}:{}",
-            self.group_id, self.artifact_id, self.latest_version,
-        )
-    }
-}
+use std::{collections::HashMap, convert::TryInto};
 
 /// Downloads the Maven master index for Google's Maven Repository
 /// and returns the XML as a String
